Ensure non-zero exit code if creating pgtap extension fails
`ON_ERROR_STOP` must be set in order for psql to return a non-zero exit code on
statement failures.

This is not much of an issue now that `IF NOT EXISTS` was added, but I thought
I'd fix it anyway given that the intent of the code is to capture errors.