MBS-12441: Handle Yandex Music links
This adds support for the Russian streaming platform
Yandex Music. They have artist, label, album and track pages,
which we are currently handling,
and genre pages, which we are not.
They also seem to have video pages which cannot be seen
without logging in. Since I cannot know what to do with them
this doesn't currently handle them at all, but doesn't
block them either so that they can still be linked as appropriate.

Since I haven't heard about TLDs affecting the content, this just
cleans all of them up to .com.