MBS-13496: Support attributes with spaces in link phrase interpolation
Using the new "a cappella" attribute was breaking link phrase
interpolation, because spaces were not accepted as part of the
interpolated "variable name".
We already had several attributes with spaces in their name
though, we just had gotten lucky we were not using any of them
in link phrases until now.

Just adding spaces to the allowed characters doesn't seem to cause
any new issues, so it seems good enough for our needs.