MBS-13328: Show Remove release edits on track artist / recording history
While other release-level edits do not usually affect the tracks
and recordings on the release's mediums, removing the release
will of course also remove the mediums, so it's equivalent
to removing all the mediums in it and it should appear in the history
for any recordings in said mediums.

Admittedly, in many cases this will lead to the recording
being subsequently removed anyway, but especially
for orphaned recordings which had relationships
and were kept behind this would be good info to have.