MBS-12439: Clean up and validate VK music links
There seems to be no limit to the optional weirdness of the VK links.
This supports /audio and /video (recording), /album (release) and
/artist links, and tries to clean up most known variations of them
to end up with something mostly consistent.
It does not try to clean up / block numeric artist IDs, which
redirect to a named artist page on desktop view (where we could
block and ask to follow the redirect) but stay numeric
in mobile view for some reason.
This means we might sometimes get duplicate numeric + named
artist pages - we'll have to trust users to be mindful when adding these.

The "unofficial" music side of VK using /wall links is not handled here
and can still be freely selected when needed.