Install po2json from git to silence warning
"DeprecationWarning: Invalid 'main' field" has been fixed for years since
https://github.com/mikeedwards/po2json/pull/97 but there was never a release.
This just installs the latest version from a git hash.

I reviewed the changelog and the diff since v0.4.1. There was one breaking
change: `format: 'jed1.x'` is now `format: 'jed'`. Other than that, I didn't
notice any issues compiling resources or browsing the website locally with
translations enabled.

As po2json appears to be abandonware, perhaps we will have to fork it
eventually, but it still works perfectly fine in the state it was left in, so I
don't think there is a need yet.