MBS-13417: Allow filtering away bootleg-only recordings
While we hide bootleg-only release groups from the overview,
and we can filter releases on status to see only non-bootlegs,
there's currently no way to hide unofficial recordings released
only on bootlegs when looking at an artist's recording page.

This filter provides this option, so that someone wanting to view
or clean up an artist's official recordings does not have
to deal with bootlegs at all unless they choose to.

Recordings that appear on both bootlegs and non-bootleg releases
are still shown, in the same way RGs with bootleg and non-bootleg
releases appear in the overview.
Standalone recordings are also still displayed.