MBS-13514: Support 6-digit label codes
For some reason label code was being restricted with a constraint
added during table generation, meaning even mirrors have it.
This is really annoying now that the org generating label codes
has started generating 6-digit ones, which we block at the
db level, turning an otherwise easy fix into a schema change.

Additionally, we don't name the constraint so postgres does automatically;
it seems the name postgres chooses is label_label_code_check.

This moves the constraint to CreateConstraints, where it won't
be added to all mirrors (if we need to change it again in the future
it will be a lot easier), and it of course amends it to actually allow
6-digit label codes.