Upgrade Flow to 0.224.0
Causing issues:
Ban accessing objects with non-literal number keys
For script and language frequencies this just means the types
need to be explicit about the possible numbers.
For MultiselectAttribute being explicit is not trivial since the ID
can be whatever, but the way we were doing it felt a bit strange
anyway. I changed it to a switch with default which seems to me
to be a much better fit for what we want anyway.

It also started detecting missing, but expected object keys for objects
with integer keys - this seems to only affect the linkPhrase test.

New feature:
Add support for number literals as keys in object literals,
e.g. const obj = {1: true, 2: false}
This means we no longer need to do the [1]: hack for the keys.
Removed the ones I found.