Add /ws/js/check-login endpoint
Quoting @amCap1712 AKA lucifer from IRC:

"[W]e intend to add a /oauth2 route to MB.org (exact path of the route yet to
be decided). [G]ateways will route requests on this path to a new MeB.org
container and that will handle OAuth stuff. It will read the cookies set by MB
and then use it to determine the current user to perform relevant actions."

"We need to do this because if we run the OAuth stuff under MeB.org, they will
need to have a MeB.org account and login there to perform any OAuth2 related
actions, which would still be fine for creation and modification of
applications but having users do that to approve access to applications would
be cumbersome."

To help resolve this, I'm exposing an endpoint, /ws/js/check-login, which can be
called by the new MeB container internally with the cookies it receives to determine
the logged-in user.

If the user is logged in, the response is as follows:

```JSON
{"id": 58244, "name": "Bitmap"}
```

Otherwise, both properties are `null`:

```JSON
{"id": null, "name": null}
```

A status code of 200 is returned in both cases.