MBS-13522: Zero-pad barcodes for ReleasesSameBarcode comparison
Without this, barcodes that are equivalent but entered with or without
zero-padding do not get seen by ReleasesSameBarcode as the same.

lpad truncates strings longer than the given length, so we need
to use greatest to make sure we allow the actual length if longer
than 14 characters.

While I use the zero-padded barcodes for comparison
(and sorting, so that they stay together on the report display)
I actually display the barcode as entered on the report,
since that seems more appropriate.