MBS-13524: Update Operabase handling for new artist URLs
Operabase now redirects their artist URLs to the form:
https://www.operabase.com/<name>-a<id>/<lang>
The id (as a<id>) is all that is needed for the URL to work,
so it is probably the best option to clean up to.
Existing URLs can, luckily, be cleaned up with relative ease.
Work and venue pages have not (yet?) changed their URLs.