Convert /edit/component templates to use Flow component syntax
For VoteCheckbox in Vote.js, Flow was unhappy with us passing
a generic ...props to the input, since in theory the props can
contain different name, type etc from what we're passing separately.
This just changes it to pass value specifically, which seems to be
the intention based on Flow typing.