Install xz through APT as it isn’t so new anymore
Back in 2017, when installed through commit 59161edb, only the version
5.1.1alpha of xz was available in Ubuntu Xenial, justifying to install
the newer version 5.2.3 from source.

Nowadays, the version 5.2.5 of xz is available in both Focal and Jammy,
so there is no need to rush at the latest version anymore, quite the
contrary given that the latest versions 5.6.x contain a backdoor.

References:

* https://packages.ubuntu.com/search?keywords=xz-utils
* https://tukaani.org/xz-backdoor/