Moving eslint config to flat config system
This is the default for eslint 9 and will be the only option
in eslint 10.

See https://eslint.org/docs/latest/use/configure/migration-guide

There's apparently a bug with the globals release we had
automatically installed and eslint flat config, so just updated
to the latest version manually since that one is fixed.