MBS-13548: obey returnto for already logged in users (#3241)
Currently if the user is already logged in, url redirects from the login page
using the returnto query parameter don't work.

This breaks certain workflows in the new OAuth testing. Fixing it by obeying
returnto for logged in users as well. If returnto is missing, fallback to
existing redirect to user's profile.