Add eslint no-implicit-coercion rule
https://eslint.org/docs/latest/rules/no-implicit-coercion

In many cases, especially when we have Flow types, it's
a lot better to just do the actual check that makes sense
based on the data instead of having boolean coercion.
For places where we don't have Flow, or coercion to
string or number, it's probably sensible to just do
the coercion explicitly.