Add eslint prefer-regex-literals rule
https://eslint.org/docs/latest/rules/prefer-regex-literals

Almost all changes are in the match section of URLCleanup.
Honestly, I'm not sure what the original reason was for that part
to be a string regex - there was a slight benefit, in that it allowed
having the regex in several lines to make it a bit more readable.
That said, there were so many cases where we were over-escaping
and, a lot worse, a bunch where we were *under-escaping*, leading
to actually incorrect regexes.
This way is a lot easier to get the escapes right.