MBS-13562 (III): Check editing unlimited strings
This patch enforces the previous definitions of both one-line string
length and indexable string size for most editable unlimited strings,
no matter whether these are currently indexed or not, in order to
prevent both MBS-13536 and MBS-13555 from happening again and also to
prevent further related issues in third-party software using this data.

The checks are made at editing time, and prevent edits to be created
when overlong strings are entered for name, sort name, artist credit
name and join phrase. An explanatory message is shown to the user,
excepted in the release editor which just shows an ISE 500 message.

The implementation comes in three parts:
* `Role::CheckOverlongString` for name (and sort name if applicable)
  of primary entities, their aliases, and medium,
* `Role::EditArtistCredit` for artist credit (name and join phrase)
  of primary entities,
* `Medium::Util` for track name and track artist credit.