Add a `read_only` flag to lib/Sql.pm
This can be used to indicate that all transactions on the Sql handle should be
read-only.

`SET TRANSACTION READ ONLY` has been moved to `Sql::begin`, because it may be
called to initialize a transaction manually. (It's currently in
`_auto_transaction`; but note that `begin` is called right above the previous
location of the code, so it's ultimately being called in the same order as
before.)