Add a `has_connector` predicate to `Server::Context`
A request may not initialize a connector (which is built lazily) if it doesn't
access the database -- or, when a read-only connector is added later, if it
only accesses one of the two connectors. A statement like
`$ctx->connector->disconnect` actually has the effect of initializing a
connector if it doesn't exist, only to immediately disconnect it. This can be
prevented by using a predicate to check if the attribute was set.