Default `READONLY` database to `READWRITE`
The `read_only` flag is now always set on the connector if the `READONLY`
database is accessed, so it's fine to use the `READWRITE` connection info if no
`READONLY` database is otherwise defined.

We've already added an `exists` method to `DatabaseConnectionFactory` (in a
previous commit) so that it can still be determined whether a `READONLY` key
was actually defined.