`Server/DatabaseConnectionFactory`: Add a `read_only` flag to `get_connector`
This is basically used to force (and verify) that the requested connector is
read-only.

Prior to this commit, the connector is only set to read-only if:

 * The `READONLY` or `PROD_STANDBY` database is requested.
 * The database has the `read_only` flag set (e.g., in `DBDefs`).

In a future commit, we'll add a read-only connector which'll default to calling
`get_connector` with `READONLY`. However, a different database than `READONLY`
could be used instead (and in fact this will happen if the `mb-set-database`
header is used); we want a way to force read-onlyness on the `get_connector`
call regardless of which key is requested.

If there's a previously-cached connector for the same key that's not read-only,
we throw an error. This shouldn't happen (e.g., we shouldn't request the
`READWRITE` key with the `read_only` flag set) unless a `fresh` connector is
also requested, in which case it's okay.