Allow connecting to any database via `READONLY_*`
Like e9181258d51639538e16feeb0311a8bccc7a9692 did for `SYSTEM_*`, this adds a
way to connect to an existing database key with the read-only flag set. For
example, accessing `READONLY_READWRITE` will create a new connector using all
of the `READWRITE` connection details, while also setting `read_only` on the
connector.

This will actually be useful in a future commit for overriding the read-only
connector's database when the `mb-set-database` header is used, without
clobbering the writable connector, which would otherwise be cached under the
same key name.