Fix installing our xgettext-js fork using Yarn v4
For some reason, installing node modules using Node.js 20.13.0 that
brings Yarn 4.0.1 would fail with the following error message:

    xgettext-js@https://github.com/metabrainz/xgettext-js.git#commit=0301681a479c1796ff6850c9bd2a169074386d92: The remote archive doesn't match the expected checksum

It is unclear which version(s) and on which platform(s) would fail here.
(It fails in macOS and Docker Composed Ubuntu Jammy, but
it works in ArchLinux and CircleCI Dockerized Ubuntu Jammy.)
When failing, the following key was added to xgettext-js’ package.json:

    "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"

It seems that it was the only difference in the installed files, thus
affecting the checksum of the fetched package xgettext-js.

To fix this issue I have upgraded our xgettext-js fork to Yarn v4 too,
in order to avoid this and have the same checksum on all platforms.

To update the checksum in `yarn.lock` file, I have temporarily set
`checksumBehavior` key to `update` in `.yarnrc.yml` file and run `yarn`.

References:

* https://yarnpkg.com/configuration/manifest#packageManager
* https://yarnpkg.com/configuration/yarnrc#checksumBehavior