Limit `getReleaseGroupRecordings` to 5 attempts
If the configured search server is unavailable (which may be a frequent
occurrence on development servers), then the release editor will attempt to
call `getReleaseGroupRecordings` every 5 seconds for infinity. This limits it
to a maximum of 5 attempts before giving up.

The release-group recordings are used for recording suggestions, so the only
effect is that they won't be available as suggestions. This commit doesn't
change that; it's just intended to reduce pointless request spam.