MBS-13581: Check "renamed into" in ArtistsThatMayBeGroups (#3265)
"Renamed into" is meant to generally be used for groups,
but it gets used to link performance names instead of
the correct "performs as" relatively often. This makes it so that those can be found.
It can cause false positives for cases where a character was renamed,
but that should be very rare and it seems like an acceptable level
of false positives.