Add `BetaOnlyEdit` placeholder for edit types in beta testing
This is a prerequisite step for IMG-84 / PR #3150.

If you were to submit any of the new EAA edits on the beta server once it's
deployed there, and then try to load any of those edits in production (whether
directly or via an edit listing), then it would cause an ISE in production.

Pulling all of the display code for these new edit types would basically mean
having to merge a very large chunk of the event-art-archive branch, as they
depend heavily on changes to the `Data::` and `Entity::` layers.

This patch adds a placeholder component, `BetaOnlyEdit`, which can be used to
link these EAA edits to the beta server (saying that they can only be viewed
there temporarily).

In a subsequent commit I will add in the machinery to actually make use of this
component for the EAA edits specifically. It may also be useful for other new
edit types we introduce in the future.