Replace /doc page query in Perl test (#3272)
da9a19b4f01f89ab6137cee16d79ebe29180bcb4 added a test to ensure that
`/login?returnto=` is obeyed if the user is already logged in. The `returnto`
page used in the test is a `/doc/` page, which actually depends on the
production Wiki server and may fail if it's under maintenance.

This commit just changes the page to one that doesn't require any production
service to be available.