MBS-13606: Drop search parameters in BetaOnlyEdit
Since this is intended a s a direct link to the edit in question,
there would seem to be no benefit in including any search
parameters in it. This is especially silly when getting to the
edit from edit searches, since the link will currently contain
the whole edit search query parameter group.