Amend d981ba8c: Fix find_best_primary_alias call (#3283)
One of the two calls to the new subroutine `find_best_primary_alias`
wasn’t checking if the second parameter was defined, causing the
warnings below to be triggered thousands of times in website logs:

    Use of uninitialized value $lang in substr at lib/MusicBrainz/Server/Data/Utils.pm line 779.
    Use of uninitialized value $lang in string eq at lib/MusicBrainz/Server/Data/Utils.pm line 785.

This patch fixes that to follow the other call. The behavior is to skip
searching for any aliases when the user interface language is undefined.