MBS-13612: Clarify you can add several art types to a file
Currently, only the error-related chooseMessage suggests
that more than one type of artwork can be selected for
the same file. It would seem checkboxes would make it obvious
but it's not - many people still don't select, say, "spine"
for files that show both the spine and another part of the art.

The hope is that these changes, agreed with Aerozol, will make it
more obvious that it's fine to select several types.