Don't set lookup-performed for ACs of length 0
`isArtistCreditStateComplete` is used to determine whether to set the
`lookup-performed` class on the single-artist autocomplete.

Apparently you can remove all AC name rows in the UI, and this will set the
class, because `.every(...)` on an empty array always returns true.

In a later commit I'd like to make it impossible to remove all AC name rows in
the UI, but this change is still correct on its own.