MBS-13624: Fix ISE on old `EditLinkAttribute` edits
`changed_display_data` only outputs fields that have changed, so the Flow type
for this edit type was incorrect, as was the `name.new === name.old` comparison
(as `name` can be `undefined`, and if it's defined, there's no need to compare
them as it only exists if they differ).

In order to display the unmodified name where possible, load the attribute type
into the display data separately. Note that we cannot load a name for all
historical edits.