Update autocomplete items on input focus
If the input is focused, and `items` is empty but `recentItems` is not, we
should set the `updateItems` flag to true; otherwise the recent items won't be
copied into `items`, and won't be displayed.

This can happen if the autocomplete initially contains a selected item, but is
then cleared without being typed into. For example, the "X" button in the
artist credit editor will clear the autocomplete if it's the last remaining AC
name row. Without this change, the recent items will not be shown, because the
`showingRecentItems` condition in `generateItems` will not be met due to the
input value being non-empty.

Typing into the autocomplete always sets `updateItems` to true, which is why
this only happens when the autocomplete is cleared indirectly.