Always refresh recent autocomplete items
The recent items may be updated by another autocomplete for the same entity
type on the page. In that case we'd want those new items to show up in the list
for all autocompletes of the same type. Prior to this commit, we only loaded
the recent items if they weren't already loaded.

I changed `getOrFetchRecentItems` to return the cached list as-is if there's
nothing to fetch (rather than always make a copy). This means we can check if
the list actually changed before dispatching the `set-recent-items` action.