MBS-13626: Always link to correct edit note from timestamp
This was always using the index to create an anchor link,
but in many cases (such as when we are editing a note or loading
just the one note for the edit in notes-received) we don't even
know what the right index would be and were always passing 0.
This meant that the note linked was always the first for the edit
in those cases, which often was not the right one.

This creates a new edit-note/id endpoint that just redirects us
to the right edit note anchor in the appropriate edit, and which
we use instead of the anchor itself everywhere except when
we are already on the edit page.