MBS-13631: Keep trailing join phrase on release -> track AC change
The changes to the artist credit copying mechanism that keeps
ACs in sync between releases and tracks meant the system
was always keeping the original join phrase at the end of the
replaced AC section.
This makes sense in most cases, but not when the whole AC
is being replaced, in which case keeping any potential
trailing phrases from the replacement AC makes a lot more sense.