MBS-13615: Skip same-entity relationships when batch-adding
If you select a set of recordings or works and try to batch-link them to a
recording or work that is part of the selected set, the whole batch-addition
will silently fail due to an internal invariant violation.

This simply skips cases where the entities would be the same in order to avoid
the error and allow the other valid relationships to be submitted.