UpdateDatabasePrivileges.pl: Remove `--other-ro-role` defaults
This is currently broken: if an array ref is passed to `GetOptions`, then new
values are pushed onto it. Instead, we want `--other-ro-role` values to
override the defaults.

However, having defaults at all makes it impossible to specify an empty list,
so I think it's better to remove them entirely.