MBS-13646: Add extracted comments to descriptions
On the model of instrument description translation source strings,
for each description (artist type, relationship type…) to be translated,
this patch provides the corresponding name as extracted comment which is
visible to translators in Weblate and doesn’t discard the existing
translations when modified.

As the output of `msguniq` isn’t ideal, a filter script is added in
order to sort and deduplicate the extracted comments.

As Weblate is just joining extracted comments with a whitespace, the
filter script is joining these with a comma instead for readability.

Reference: https://www.gnu.org/software/gettext/manual/html_node/PO-Files.html