MBS-13645: Fix translations of edit type names
We were not actually using the contexts we added when
showing translations, so these were mostly just not working at all.

Since contexts vary per edit type, this sets a default context
of 'edit type' and allows overriding it for the few edits that
change it (adding 'singular' or 'plural').

In statistics/Edits I chose to pass l_edit_name rather than
passing the context and translating on the JS side because
the whole thing is a construct for display so we might as well
simplify it. This is the same that is already done for EditTypeList.