Remove unused {Create,Drop}MQTriggers.sql files
These were used in conjunction with the old CAA-indexer. The newer
artwork-indexer, which is now fully deployed, uses a different set of triggers.

Note: I removed a reference to caa/CreateMQTriggers.sql in HACKING-CAA.md, but
that entire section is still mostly wrong; I'll update the rest of it in a
later commit.