IMG-154 (diagram json): Add EAA tables to diagrams
The source format doesn’t allow to distinguish the table `art_type` in
the scheme `cover_art_archive` from the new table by the same name in
the scheme `event_art_archive`.

Adding it as a duplicate for now, but it will require more manual steps.

The source format needs to be improved in the longer term.