MBS-13677: Show cover art presence in indexed searches
This was loading the artwork (which we don't actually need)
but not actually updating the has_cover_art flag we actually
care about for displaying the icon.

I probably broke this when I introduced load_has_cover_art
and forgot to use it here? In any case, that's what we need here.
Should also make searches slightly faster, I expect.