MBS-13675: Split statistic events into history translation component
History events (used in the history page and the timeline)
are completely different from most statistics, and it seems
sensible to translate them separately. In many cases, they
probably are even less important than the rest of the
statistics strings.

This first commit makes the necessary code changes
to move the translations to a new history component.
In next commits I'll generate the relevant .pot file,
and generate start .po files to save the currently translated lines.