MBS-13528: Consider edit note empty if it's Hangul fillers
HANGUL FILLER is a character often used to have "empty usernames"
and similar things since it's an empty space but not considered
a space char. Some user was using it as an edit note
to add releases, which is not nice - this considers it, together
with other similar Hangul fillers, insufficient for a valid note.