Replace blog post with MBS version in tag message
Git tag message used to link to the corresponding blog post URL.
However such URL depends on the date the blog post is published.
As a consequence, the tag could be set at the last minute only.

This is more of an issue now that the companion MusicBrainz Docker
Compose project relies on this tag for prebuilding a Docker image.

This patch replaces the blog post URL with the MBS version URL in Jira.
Such URL just uses a serial number, and can be retrieved via a REST API.
Version’s description can be set manually to the blog post URL later on.

It assumes that the version in Jira is marked as released after tagging.
The Git tag can now be set before building Docker images.
It also paves the way for further automation.