MBS-13453: Avoid double-decoding some errors
While the idea in 20640deaa0841d7d55533dc3e7c8a8cbe2722fc1
was correct, the change was too wide in application;
it was also decoding errors that did not need it, causing
a wide character ISE because of the double decoding.

The double-decoding could be triggered (before this patch)
by doing a utf8 change on a sql data file for a test and then
running it - I changed work.sql to have an invalid work ID
(to trigger an error) and ♥ instead of Test as a work name.
Then I just triggered the error with
prove -lv t/tests.t :: --tests Edit::Work::Create

mwiencek looked into the root causes and said:
"We do always get a UTF-8 string. But sometimes it's just a raw
byte string (with no UTF8 flag set), and other times it's already
been decoded (UTF8 flag on)."

So this checks whether the flag is on, and only decodes the string
if it is not.

Co-Authored-By: Michael Wiencek <mwtuea@gmail.com>