MBS-13682: Add language filter to Works tab
This allows filtering an artist's works by the lyrics language.
Since giving all languages as options would be overwhelming
and not very useful, this looks at what languages are actually used
and offers any of these or [not set] for works with no languages set
(since yvanzo worries [none], which we use for other filters,
will be too confusing for work lyrics when [No lyrics] exists as well).