MBS-13579: Preserve existing release labels duplicated by seeded ones (#3260)
Prior to this commit, if you seeded a release label to an existing release that
duplicated an existing release label on that release, the release editor would
throw an error upon submission, as it would be trying to submit data that
already exists.

This patch preserves the existing release label (and its database row ID),
discarding the seeded one. Note that a "release label" is a label and catalog
number pair, meaning the catalog number is included for de-duplication
purposes.

The ticket presents three use cases. The first one is still possible, the
second one is still impossible, and the third one is now possible. An
alternative implementation would be to concatenate the seeded release labels
with the existing ones (discarding any duplicates, like here), and adding a
`removed` field to allow removing existing release labels. However, you'd need
to know the label IDs (and associated catalog numbers) of the release labels
you'd like to remove in advance, which isn't currently a requirement: it
becomes harder to replace all of the existing release labels with a new set.

There is one additional feature this commit implements: if you seed a release
label, and there already exists one on the release with the same label (but
different catalog number), we will edit the existing release label with the new
catalog number. The same is true if the catalog number matches but not the
label (but label matches are always preferred). An exception to this is if the
existing release label is being preserved by the seeded parameters.