MBS-8217 (II): Don't show guess case preview while dragging
Make the FormRowNameWithGuessCase and
FormRowSortNameWithGuessCase components, the GuessCase
control, and the tracklist editor code avoid showing
previews if a button is held while the mouse enters a
"guess case" or "guess sort name" button.

Otherwise, it's easy to clear the user's selection if
they accidentally mouse over the button while dragging
to select text.