MBS-13696: Check for invalid notes more in ws/js
We are now checking for invalid edit notes everywhere except
in two places: when editing (as opposed to adding) a release,
and in the relationship editor,
both of which edit via ws/js.
There seems to be no good reason not to expand this check
to these last places.
Since the relationship editor doesn't process anything and
just uses the string passed to detach_with_error, this changes
the error to the full translated string.