MBS-13700: Only open edit note links in new tab if needed
We decided to open edit note links in new tabs to avoid users
accidentally navigating away from their unsent notes and votes
but we were doing that everywhere except from
the single edit pages.
This is overkill - the only place we really need to do this
is ListEdit.
This separates isOnEditPage and a new flag isOnEditList, and opens
anything not on an edit list in the same tab.