MBS-13707: Calculate needed columns first in getRelationshipRows
The first row(s) weren't getting all needed columns after 7072352ad4084e6bd21ef6249520c276998f7737
removed the nested component, which means it probably only worked
because of things getting re-rendered.

The actual right way to do this would seem to be to calculate it once per relationship group, since we want it to apply
to the whole group anyway.