Completely avoid conflicting keys in `VarArgs.prototype.getKey`
The `getKey` method added in 6f57c167a3996f35cb1012d05f93e709f66cbdd6 doesn't
prevent conflicts in cases like this example (pulled from externalLinks.js):

    typeDescription = exp.l('{description} ({url|more documentation})', {
      description: expand2react(l_relationships(linkType.description)),
      url: '/relationship/' + linkType.gid,
    });

This is because the inner `expand2react` call does not share a `VarArgs`
instance with the `exp.l` call, so separate `nextKey` counters are used. Thus
the first link in the `description` will get the key "a-1", and so will the
link generated for `url`.

We can avoid this by just using a global counter for `nextKey`. This is not
entirely ideal, because changing the key on every render prevents React's
reconciliation algorithm from reusing the same elements. I'd like to submit a
better fix soon, but it's much more involved. It's also not too big of a deal,
because the elements that are typically interpolated are small (like links),
not giant subtrees that would benefit more from consistent keys.