Log plackup response time
This copies the "combined" plackup log format and appends the value of the
`X-Runtime` header which is injected by `Plack::Middleware::Runtime`.

Note that the modified log format is only used in production. (The `AccessLog`
middleware is enabled by default in development, though I didn't find out how
to change the log format used in that case.)