Fix runit signal propagation
Signal propagation didn't work currently where `exec sudo` was used. This meant
issuing a command like `sv down service` wouldn't be able to stop the service.
I reviewed the "Signal handling" section of the sudo manpage, but still didn't
quite understand why this behavior occurs.

The solution is to use `chpst`, which comes with runit and works perfectly
fine. The only change needed is to manually set `$HOME`, which was previously
set via sudo's `-H` flag. (Not every service uses `$HOME`, but I set it
everywhere just to be safe.)