Upgrade Chrome in tests to 127.0.6533.88
This upgrade caused any test that used `handleAlert` to fail. IIUC, this was
caused by [1], which is an intentional change to no longer show `beforeunload`
alerts at all when the WebDriver is in use.

It's still useful to be able to test that an alert would've been shown, as it's
the only way of verifying that such functionality works to prevent data loss.
I've replaced `handleAlert` with `assertBeforeUnloadAlertWasShown`. The main
difference is that we aren't actually handling the alert (as none are
triggered), but are checking that it would've been shown in a normal browser
environment.

[1] https://chromium.googlesource.com/chromium/src/+/b14c93608871784e41d6d40f1c5952cf24aa39db