Fix runit service logging
The "current" log files produced by svlogd were completely empty in many cases.
This is because we weren't redirecting stderr to stdout in the service "run"
files.

This only affected services within the musicbrainz-tests container, which
matters because we upload these log files as build artifacts.
(start_template_renderer.sh is also used by the production images, but we don't
use svlogd there.)