Revert script/database_exists to fallback behavior
Prior to e61190f6e26ee73d45becfdd9ea0d58b196f533b, the database_exists script
did not require e.g. `MAINTENANCE` to be defined if `READWRITE` was (because
`get` with `MAINTENANCE` falls back to `READWRITE`). There was no real reason
to change this script to use the new `exists` method over `get`, and in fact
some scripts in musicbrainz-docker rely on the old behavior, so I'm reverting
the change.