New way of assigning React keys to variables substitutions
In 48006218adf670d1793a8ed2738fd5b417a868e7 it was said that the solution there
was not ideal, as it would cause React to replace elements that could otherwise
be reused.

In fact this has resulted in issues like MBS-13729, where userscripts had
previously expected certain elements to remain stable.

The new solution here is to always wrap React elements used as variable
substitution arguments with a fragment using the substitution key as the React
key. For example:

    exp.l('{a}', {a: <span />})

Will result in:

    <React.Fragment key="a">
      <span />
    </React.Fragment>

If the substitution value for `a` is an array of React elements, each element
will be wrapped in a fragment using an indexed key, like `a-0`, `a-1`, etc.

I've confirmed that this fixes MBS-13729, and doesn't regress on the example
provided in 48006218adf670d1793a8ed2738fd5b417a868e7.