MBS-13729: Handle missing `medium_positions` on release merge
The error described in MBS-13729 was ultimately caused by a userscript removing
required fields from the page: in particular, the "merge strategy" selection
was being removed, which causes the page to reload with an error ("Merge
strategy field is required").

The script also removes all of the fields contained in `medium_positions`.
According to our Flow types, this is unexpected. We expect there to be a
corresponding field in `medium_positions` for each medium.

To address this, I've first modified the Flow type to specify that certain
indices within `medium_positions` may be `void`. Next, I've added some fallback
code to construct a `medium_positions` subfield from its corresponding medium
data if the former is absent.

Note that cf55fc88af3b3ce83c6d5a2969eeeabee031c790 already unbreaks the script
in question, but this ensures that the merge page presents the error gracefully
to the user rather than as an exception stack trace.