Convert `dialog_result` part of root/forms/dialog.tt to React
When an entity creation form is submitted within an `AddEntityDialog`
(accessible from an autocomplete), the submitted iframe will reload with a
blank page containing the just-created entity JSON in a global variable named
`dialogResult`. The parent document retrieves `dialogResult` from the child
iframe and uses it to populate the autocomplete.

(Note that `dialogResult` comes directly from the server and can't be injected
with arbitrary values; see `Controller::Role::Create`.)

This "blank" page was toggled via a conditional in root/forms/dialog.tt. It's
been ported to React in a separate file, root/forms/DialogResult.js, and the
conditional has been moved from the template to the server.